Move the validate hook to the new command class.
