+++ src/Commands/config/ConfigImportCommands.php
@@ -2,6 +2,8 @@
 
 namespace Drush\Commands\config;
 
+use Consolidation\AnnotatedCommand\CommandData;
+use Consolidation\AnnotatedCommand\CommandError;
 use Drush\Commands\DrushCommands;
 
 /**
@@ -43,5 +45,31 @@ class ConfigImportCommands extends DrushCommands
         /** @var \Drush\Drupal\Commands\config\ConfigImportCommands $config_import_commands_service */
         $config_import_commands_service = \Drupal::service('config.import.commands');
         return $config_import_commands_service->import($label, $options);
+    }
+
+    /**
+     * @hook validate config-import
+     * @param \Consolidation\AnnotatedCommand\CommandData $commandData
+     * @return \Consolidation\AnnotatedCommand\CommandError|null
+     */
+    public function validate(CommandData $commandData)
+    {
+        $msgs = [];
+        if ($commandData->input()->getOption('partial') && !\Drupal::moduleHandler()->moduleExists('config')) {
+            $msgs[] = 'Enable the config module in order to use the --partial option.';
+        }
+
+        if ($source = $commandData->input()->getOption('source')) {
+            if (!file_exists($source)) {
+                $msgs[] = 'The source directory does not exist.';
+            }
+            if (!is_dir($source)) {
+                $msgs[] = 'The source is not a directory.';
+            }
+        }
+
+        if ($msgs) {
+            return new CommandError(implode(' ', $msgs));
+        }
     }
 }
